update the type of l2_block_hash in the L1L2BlockMapping
